refactor: setup for standardizing benchmarks
