+++ Makefile
@@ -1,4 +1,5 @@
-all    :; dapp build
-clean  :; dapp clean
-test   :; dapp test --rpc-url ${ETH_RPC_URL}
-deploy :; dapp create Convex
+benchmark-all       :; bash scripts/benchmark-all.sh
+benchmark-dapptools :; bash scripts/benchmark-dapptools.sh
+benchmark-foundry   :; bash scripts/benchmark-foundry.sh
+benchmark-ganache   :; bash scripts/benchmark-ganache.sh
+benchmark-hardhat   :; bash scripts/benchmark-hardhat.sh
