+++ scripts/benchmark-all.sh
@@ -1,0 +1,11 @@
+echo -e "\n--- BENCHMARKING DAPPTOOLS ---"
+bash scripts/benchmark-dapptools.sh
+
+echo -e "\n--- BENCHMARKING FOUNDRY ---"
+bash scripts/benchmark-foundry.sh
+
+echo -e "\n--- BENCHMARKING GANACHE ---"
+bash scripts/benchmark-ganache.sh
+
+echo -e "\n--- BENCHMARKING HARDHAT ---"
+bash scripts/benchmark-hardhat.sh
