+++ scripts/benchmark-dapptools.sh
@@ -1,0 +1,5 @@
+. ./.env
+
+# dapptools has no option to set gas limit, but it's hardcoded by default
+# anyway so shouldn't be make a difference
+time dapp test --rpc-url $ETH_RPC_URL --rpc-block $FORK_BLOCK
