+++ scripts/benchmark-foundry.sh
@@ -1,0 +1,3 @@
+. ./.env
+
+time forge test --rpc-url $ETH_RPC_URL --fork-block-number $FORK_BLOCK --gas-limit $GAS_LIMIT
