+++ scripts/benchmark-ganache.sh
@@ -1,0 +1,4 @@
+. ./.env
+
+# Uses Ganache v6 (@beta)
+time yarn ts-node ./scripts/convex.ganache.ts
