+++ scripts/benchmark-hardhat.sh
@@ -1,0 +1,3 @@
+. ./.env
+
+time TS_NODE_FILES=true yarn ts-node ./scripts/convex.hardhat.ts
